+++ src/main/java/com/cookeep/cookeep/domain/verification/application/verification/VerificationSender.java
@@ -1,4 +1,4 @@
-package com.cookeep.cookeep.domain.verification.application.sms;
+package com.cookeep.cookeep.domain.verification.application.verification;
 
 public interface VerificationSender {
 	void send(String to, String title, String content);
